+++ lib/renderers/dom.js
@@ -1,7 +1,7 @@
 import 'babel-polyfill';
 import React from 'react';
 import { hydrate } from 'react-dom';
-import App from './App';
+import App from '../components/App';
 
 hydrate(
   <App />,
