+++ lib/renderers/server.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOMServer from 'react-dom/server';
-import App from './components/App';
+import App from '../components/App';
 
 const serverRender = () => {
   return ReactDOMServer.renderToString(
